fix(aggregate,scalar)!: compute f32 embeddings in f64 to end the cancellation defect class
Three consecutive review rounds found the same numerical class inside the
previous round's fix, because the contract, not the arithmetic, was wrong: an
f32 embedding accumulated in f32, so every power-of-two scaling scheme had a
next adversarial counterexample (subnormal flush, exact-cancellation
fabrication, mixed-magnitude EMA collapse).

Set `f32`'s associated compute type to `f64`. `f32` becomes a storage-only
scalar and is no longer a `Real`; `f64` is the sole compute domain, and both
shipped scalars (and the int8 test double) compute there. Every f32 is exact in
f64, every f32 subnormal is a normal f64, and f32::MAX squared sits far inside
f64's range, so the entire per-fold power-of-two shift apparatus becomes
unnecessary and is deleted: accumulation_shift, ema_shift, saliency_shift,
scaled_l2_norm, and Real's MAX_EXP/MIN_EXP all go.

Two hazards survive into f64 itself and are handled where they arise:

- Cancellation across a wide exponent spread now uses compensated (Neumaier)
  summation, so an exactly cancelling weighted sum lands at exactly zero under
  every association order and NonFinite keeps meaning "no direction".
- A norm unrepresentable although the vector is finite is still handled by the
  scale-aware renormalization, which never forms that norm.

The one irreducible ceiling is SaliencyWeighted, which squares magnitudes: its
f64 window is sqrt(MIN_POSITIVE)..sqrt(MAX) (~1.5e-162..~1.3e154), outside which
it returns NonFinite rather than rescale with the fabrication-prone shift that
widening to f64 exists to retire. Every realistic magnitude is far inside it.

The policy trait's compute-scalar parameter now defaults to f64 (the domain
every ordinary embedding computes in); the bare `dyn AggregatePolicy` spelling
and object safety are unchanged. AggregatePolicyKind and its serde wire format
are untouched — configuration stays f32 and widens at use.

Also sweep every growing collection on a Result-returning path in aggregate for
fallibility (Codex R4 MEDIUM): the coverage/borrow/widen buffers in `aggregate`,
the accumulator and compensation in weighted_sum_renorm, EMA's state, and
saliency's weights now grow through try_reserve and map a refused allocation to
WinditError::AllocFailed. A new tests/aggregate_alloc.rs arms a refusing global
allocator and proves each built-in policy reports AllocFailed rather than
aborting.

The f32 bit-exactness guard, which pinned bit-identity with naive f32 folding,
is replaced: accumulating in f64 and rounding once is deliberately more accurate
than an f32 fold, so the guard now asserts the scale-aware result is within one
ulp of the naive f64 fold and L2-unit to f64 precision.

+++ src/test_support.rs
@@ -3,10 +3,11 @@
 //! Two embedding doubles and one float-closeness assertion ([`assert_close`])
 //! serve every module's `tests.rs`, so the crate carries its test [`Vector`]
 //! implementations in one place instead of redefining them per module.
-//! [`TestVec`] is the ordinary f32 embedding; [`TestQuantVec`] stores a narrower
-//! scalar than it computes in, which is the only way to reach `aggregate`'s
-//! widening path. No helper needs `std`: they normalize through `libm::sqrtf`
-//! and avoid the std-only float methods.
+//! [`TestVec`] is the ordinary f32 embedding — which now computes in `f64`, so
+//! its `from_unnormalized` takes `&[f64]` and narrows on the way into storage;
+//! [`TestQuantVec`] stores an integer scalar narrower still. No helper needs
+//! `std`: they normalize through `libm::sqrt` and avoid the std-only float
+//! methods.
 
 use std::vec::Vec;
 
@@ -18,8 +19,10 @@ use crate::{
 
 /// A minimal embedding double that L2-normalizes on construction.
 ///
-/// `from_unnormalized` divides by the L2 norm (via `libm::sqrtf`), rejecting an
-/// empty slice with [`WinditError::Empty`] and a zero or non-finite norm with
+/// It stores `f32` but computes in `f64` (`f32`'s compute type), so
+/// `from_unnormalized` takes `&[f64]`, normalizes there, and narrows each
+/// component to `f32` on the way into storage. It rejects an empty slice with
+/// [`WinditError::Empty`] and a zero or non-finite norm with
 /// [`WinditError::NonFinite`].
 pub(crate) struct TestVec(pub(crate) Vec<f32>);
 
@@ -30,15 +33,15 @@ impl Vector for TestVec {
     &self.0
   }
 
-  fn from_unnormalized(v: &[f32]) -> Result<Self, WinditError> {
+  fn from_unnormalized(v: &[f64]) -> Result<Self, WinditError> {
     if v.is_empty() {
       return Err(WinditError::Empty);
     }
-    let norm = libm::sqrtf(v.iter().map(|x| x * x).sum::<f32>());
+    let norm = libm::sqrt(v.iter().map(|x| x * x).sum::<f64>());
     if !norm.is_finite() || norm == 0.0 {
       return Err(WinditError::NonFinite);
     }
-    Ok(Self(v.iter().map(|x| x / norm).collect()))
+    Ok(Self(v.iter().map(|x| (x / norm) as f32).collect()))
   }
 }
 
@@ -64,14 +67,14 @@ impl Vector for TestQuantVec {
     if v.is_empty() {
       return Err(WinditError::Empty);
     }
-    let norm = libm::sqrtf(v.iter().map(|x| x * x).sum::<f32>());
+    let norm = libm::sqrt(v.iter().map(|x| x * x).sum::<f64>());
     if !norm.is_finite() || norm == 0.0 {
       return Err(WinditError::NonFinite);
     }
     Ok(Self(
       v.iter()
         .map(|x| {
-          let scaled = libm::roundf(x / norm * 127.0);
+          let scaled = libm::round(x / norm * 127.0);
           TestQuant(scaled.clamp(-127.0, 127.0) as i8)
         })
         .collect(),
